refactor: extract user api calls into api/users.ts
